[#33] FeedDetailView DI 연결 및 코드 정리
